add buyable and sellable interfaces to class

See merge request fit2099/fit2099-s2-2023/assignment-groups/CL_AppliedSession9_Group1/project!6

+++ src/game/actors/traders/pricings/IncreasedPricing.java
@@ -34,7 +34,7 @@ public class IncreasedPricing implements Pricing {
     public int getPrice(int originalPrice){
         Random rand = new Random();
         if (rand.nextInt(100) <= this.chanceToIncrease) {
-            return originalPrice * (1+(this.percentageIncrease/100));
+            return (int)(originalPrice * (1+(this.percentageIncrease/100.0)));
         } else {
             return originalPrice;
         }
